+++ back/.env
@@ -1,2 +1,1 @@
-PRIVATE_KEY=2ba827386cd079fc2cab1851c96773b3daf3d927a502fa454aeec0f58fb4bf83
-NODE_ENV=dev
+PRIVATE_KEY=
