+++ back/test/ecrecover.test.ts
@@ -6,7 +6,5 @@ const sig = '0x67f8a1e8e52a8401c7bf21161637b381f564a424b615bc741575518f3df051dc4
 const address = '0x945b59d5a86e2c7d51f961f4987e8b0cad4a4f57'
 
 describe('ecrecover', () => {
-  test('recovers account', () => {
-    expect(ecrecover(message, sig)).toEqual(address)
-  })
+  test('recovers account', () => expect(ecrecover(message, sig)).toEqual(address))
 })
