+++ back/test/utils.ts
@@ -1,8 +1,11 @@
 export const account = '0x87eb390df1e05ef0560e387206f5997034cd6f28'
 export const did = 'did:ethr:rsk:0x87eb390df1e05ef0560e387206f5997034cd6f28'
+
 export const type = 'Test'
 export const subject = 'Claim!'
+
 export const code = 'CODE'
+
 export const jwt = 'jwt'
 
 export const issuerPrivateKey = '2ba827386cd079fc2cab1851c96773b3daf3d927a502fa454aeec0f58fb4bf83' // do not use
